requirements.txt
python-telegram-bot>=20.0
openai>=1.0.0
chromadb==0.4.22
pydantic<2.0.0,>=1.10.0
python-dotenv>=0.19.0
tiktoken>=0.5.0
numpy<2.0.0
fastapi<0.100.0
google-play-scraper
